+++ src/main/java/dev/lambdaurora/lambdynlights/engine/source/EntityDynamicLightSource.java
@@ -7,20 +7,26 @@
  * see the LICENSE file.
  */
 
-package dev.lambdaurora.lambdynlights.engine;
+package dev.lambdaurora.lambdynlights.engine.source;
 
 import dev.lambdaurora.lambdynlights.LambDynLights;
+import dev.lambdaurora.lambdynlights.engine.DynamicLightingEngine;
+import dev.lambdaurora.lambdynlights.engine.lookup.SpatialLookupEntry;
+import dev.lambdaurora.lambdynlights.engine.lookup.SpatialLookupEntityEntry;
+import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.Range;
 
+import java.util.stream.Stream;
+
 /**
- * Represents a dynamic light source.
+ * Represents an entity-based dynamic light source.
  *
  * @author LambdAurora
  * @version 4.0.0
  * @since 1.0.0
  */
-public interface DynamicLightSource {
+public interface EntityDynamicLightSource extends DynamicLightSource {
 	/**
 	 * {@return the dynamic light source X-coordinate}
 	 */
@@ -61,4 +67,16 @@ public interface DynamicLightSource {
 	 * Executed at each tick.
 	 */
 	void dynamicLightTick();
+
+	@Override
+	default Stream<SpatialLookupEntry> splitIntoDynamicLightEntries() {
+		int x = MathHelper.floor(this.getDynamicLightX());
+		int y = MathHelper.floor(this.getDynamicLightY());
+		int z = MathHelper.floor(this.getDynamicLightZ());
+
+		int cellKey = DynamicLightingEngine.hashAt(x, y, z);
+		SpatialLookupEntry entry = new SpatialLookupEntityEntry(cellKey, this);
+
+		return Stream.of(entry);
+	}
 }