+++ src/main/java/dev/lambdaurora/lambdynlights/engine/source/EntityDynamicLightSourceBehavior.java
@@ -7,19 +7,18 @@
  * see the LICENSE file.
  */
 
-package dev.lambdaurora.lambdynlights.engine;
+package dev.lambdaurora.lambdynlights.engine.source;
 
 import dev.lambdaurora.lambdynlights.LambDynLights;
+import dev.lambdaurora.lambdynlights.engine.DynamicLightingEngine;
 import it.unimi.dsi.fastutil.longs.LongOpenHashSet;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.LevelRenderer;
+import it.unimi.dsi.fastutil.longs.LongSet;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.ChunkSectionPos;
 import net.minecraft.core.Direction;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.entity.Entity;
 import org.jetbrains.annotations.ApiStatus;
-import org.jetbrains.annotations.NotNull;
 
 /**
  * Represents the behavior of a dynamic light source.
@@ -29,7 +28,7 @@ import org.jetbrains.annotations.NotNull;
  * @since 4.0.0
  */
 @ApiStatus.Internal
-public interface DynamicLightSourceBehavior extends DynamicLightSource {
+public interface EntityDynamicLightSourceBehavior extends EntityDynamicLightSource {
 	/**
 	 * {@return the dynamic light source previous X-coordinate}
 	 */
@@ -83,7 +82,9 @@ public interface DynamicLightSourceBehavior extends DynamicLightSource {
 	 * @param entity the entity to tick
 	 */
 	static void tickEntity(Entity entity) {
-		var lightSource = (DynamicLightSourceBehavior) entity;
+		if (!LambDynLights.get().shouldTick()) return;
+
+		var lightSource = (EntityDynamicLightSourceBehavior) entity;
 
 		if (entity.isRemoved()) {
 			lightSource.setDynamicLightEnabled(false);
@@ -97,9 +98,11 @@ public interface DynamicLightSourceBehavior extends DynamicLightSource {
 		}
 	}
 
-	default boolean lambdynlights$updateDynamicLight(@NotNull LevelRenderer renderer) {
-		if (!this.shouldUpdateDynamicLight())
-			return false;
+	default LongSet getDynamicLightChunksToRebuild(boolean forced) {
+		if (forced) {
+			return this.lambdynlights$getTrackedLitChunkPos();
+		}
+
 		double x = this.getDynamicLightX();
 		double y = this.getDynamicLightY();
 		double z = this.getDynamicLightZ();
@@ -109,61 +112,55 @@ public interface DynamicLightSourceBehavior extends DynamicLightSource {
 
 		int luminance = this.getLuminance();
 
-		if (Math.abs(deltaX) > 0.1D || Math.abs(deltaY) > 0.1D || Math.abs(deltaZ) > 0.1D || luminance != this.getLastDynamicLuminance()) {
-			this.updateDynamicLightPreviousCoordinates();
-			this.setLastDynamicLuminance(luminance);
-
-			var newPos = new LongOpenHashSet();
-
-			if (luminance > 0) {
-				var chunkPos = new BlockPos.Mutable(
-						ChunkSectionPos.blockToSectionCoord(x),
-						ChunkSectionPos.blockToSectionCoord(y),
-						ChunkSectionPos.blockToSectionCoord(z)
-				);
-
-				LambDynLights.scheduleChunkRebuild(renderer, chunkPos);
-				LambDynLights.updateTrackedChunks(chunkPos, this.lambdynlights$getTrackedLitChunkPos(), newPos);
-
-				var directionX = (MathHelper.floor(x) & 15) >= 8 ? Direction.EAST : Direction.WEST;
-				var directionY = (MathHelper.floor(y) & 15) >= 8 ? Direction.UP : Direction.DOWN;
-				var directionZ = (MathHelper.floor(z) & 15) >= 8 ? Direction.SOUTH : Direction.NORTH;
-
-				for (int i = 0; i < 7; i++) {
-					if (i % 4 == 0) {
-						chunkPos.move(directionX); // X
-					} else if (i % 4 == 1) {
-						chunkPos.move(directionZ); // XZ
-					} else if (i % 4 == 2) {
-						chunkPos.move(directionX.getOpposite()); // Z
-					} else {
-						chunkPos.move(directionZ.getOpposite()); // origin
-						chunkPos.move(directionY); // Y
-					}
-					LambDynLights.scheduleChunkRebuild(renderer, chunkPos);
-					LambDynLights.updateTrackedChunks(chunkPos, this.lambdynlights$getTrackedLitChunkPos(), newPos);
-				}
-			}
-
-			// Schedules the rebuild of removed chunks.
-			this.lambdynlights$scheduleTrackedChunksRebuild(renderer);
-			// Update tracked lit chunks.
-			this.lambdynlights$setTrackedLitChunkPos(newPos);
-			return true;
+		if (Math.abs(deltaX) <= 0.1 && Math.abs(deltaY) <= 0.1 && Math.abs(deltaZ) <= 0.1 && luminance == this.getLastDynamicLuminance()) {
+			return LongSet.of();
 		}
-		return false;
+
+		var newPos = new LongOpenHashSet();
+
+		if (luminance > 0) {
+			this.gatherClosestChunks(newPos, x, y, z);
+		}
+
+		var result = new LongOpenHashSet(newPos);
+		result.addAll(this.lambdynlights$getTrackedLitChunkPos());
+
+		this.updateDynamicLightPreviousCoordinates();
+		this.setLastDynamicLuminance(luminance);
+		this.lambdynlights$setTrackedLitChunkPos(newPos);
+
+		return result;
 	}
 
-	LongOpenHashSet lambdynlights$getTrackedLitChunkPos();
+	private void gatherClosestChunks(LongSet chunks, double x, double y, double z) {
+		var chunkPos = new BlockPos.Mutable(
+				ChunkSectionPos.blockToSectionCoord(x),
+				ChunkSectionPos.blockToSectionCoord(y),
+				ChunkSectionPos.blockToSectionCoord(z)
+		);
 
-	void lambdynlights$setTrackedLitChunkPos(LongOpenHashSet trackedLitChunkPos);
+		chunks.add(ChunkSectionPos.asLong(chunkPos.getX(), chunkPos.getY(), chunkPos.getZ()));
 
-	@SuppressWarnings("resource")
-	default void lambdynlights$scheduleTrackedChunksRebuild(@NotNull LevelRenderer renderer) {
-		if (Minecraft.getInstance().level == this.dynamicLightWorld()) {
-			for (long pos : this.lambdynlights$getTrackedLitChunkPos()) {
-				LambDynLights.scheduleChunkRebuild(renderer, pos);
+		var directionX = (MathHelper.floor(x) & 15) >= 8 ? Direction.EAST : Direction.WEST;
+		var directionY = (MathHelper.floor(y) & 15) >= 8 ? Direction.UP : Direction.DOWN;
+		var directionZ = (MathHelper.floor(z) & 15) >= 8 ? Direction.SOUTH : Direction.NORTH;
+
+		for (int i = 0; i < 7; i++) {
+			if (i % 4 == 0) {
+				chunkPos.move(directionX); // X
+			} else if (i % 4 == 1) {
+				chunkPos.move(directionZ); // XZ
+			} else if (i % 4 == 2) {
+				chunkPos.move(directionX.getOpposite()); // Z
+			} else {
+				chunkPos.move(directionZ.getOpposite()); // origin
+				chunkPos.move(directionY); // Y
 			}
+			chunks.add(ChunkSectionPos.asLong(chunkPos.getX(), chunkPos.getY(), chunkPos.getZ()));
 		}
 	}
+
+	LongSet lambdynlights$getTrackedLitChunkPos();
+
+	void lambdynlights$setTrackedLitChunkPos(LongSet trackedLitChunkPos);
 }