+++ src/main/java/dev/lambdaurora/lambdynlights/mixin/DebugScreenOverlayMixin.java
@@ -25,16 +25,18 @@ import java.util.List;
  * Adds a debug string for dynamic light sources tracking and updates.
  *
  * @author LambdAurora
- * @version 3.2.2
+ * @version 4.0.0
  * @since 1.3.2
  */
 @Mixin(DebugScreenOverlay.class)
 public class DebugScreenOverlayMixin {
 	@Inject(method = "getGameInformation", at = @At("RETURN"))
 	private void onGetLeftText(CallbackInfoReturnable<List<String>> cir) {
+		var prefix = TextFormatting.LIGHT_PURPLE + "[LDL] " + TextFormatting.RESET;
+
 		var list = cir.getReturnValue();
 		var ldl = LambDynLights.get();
-		var builder = new StringBuilder("Dynamic Light Sources: ");
+		var builder = new StringBuilder(prefix + "Dynamic Light Sources: ");
 		builder.append(ldl.getLightSourcesCount())
 				.append("/")
 				.append(DynamicLightingEngine.MAX_LIGHT_SOURCES)
@@ -50,6 +52,9 @@ public class DebugScreenOverlayMixin {
 
 		builder.append(')');
 		list.add(builder.toString());
+
+		list.add(prefix + "Compute Spatial Lookup Timing: %.3fms (avg. 40t)"
+				.formatted(ldl.engine.getComputeSpatialLookupTime() / 1_000_000.f));
 
 		if (LambDynLightsConstants.isDevMode()) {
 			list.add(TextFormatting.RED + LambDynLightsConstants.DEV_MODE_OVERLAY_TEXT);