+++ CHANGELOG.md
@@ -239,7 +239,11 @@
 - Added display entities dynamic lighting ([#209](https://github.com/LambdAurora/LambDynamicLights/issues/209)).
   - This affects both block and item displays.
   - If a custom brightness is defined then dynamic lighting disables itself.
+- Added debug settings and renderers to facilitate debugging.
+  - Added a debug renderer to active dynamic lighting cells.
+  - Added a debug renderer to display chunk rebuilds.
 - Refactored heavily the ticking of entity dynamic lighting to make it more abstract.
+- Refactored heavily how chunk rebuilds are queued and how dynamic light sources are represented.
 - Fixed API publication for loom-based setups.
 
 [SpruceUI]: https://github.com/LambdAurora/SpruceUI "SpruceUI page"